requirements.txt
tensorflow==2.0.0
tensorflow-serving-api==2.0.0
opencv-python==4.1.1.26
hdrhistogram
numpy
requests